refactor: decompose operations frontend
