+++ CLAUDE.md
@@ -1,0 +1,5 @@
+# Benson ERP Claude instructions
+
+- Follow `AGENTS.md` and the nearest nested `AGENTS.md` for the file being changed.
+- For `benson-app/web/`, the binding frontend boundaries and verification commands are in `benson-app/web/AGENTS.md`.
+- Preserve authenticated deny-by-default access, server-side authorization, guarded lead mutations, immutable audit history, and private attachment handling while refactoring.
