+++ benson-app/web/eslint.config.js
@@ -9,5 +9,9 @@ export default defineConfig([
   {
     files: ["src/**/*.{ts,tsx}"],
     languageOptions: { globals: { fetch: "readonly", document: "readonly", window: "readonly", Intl: "readonly" } },
+    rules: {
+      "max-lines": ["error", { max: 350, skipBlankLines: true, skipComments: true }],
+      "max-lines-per-function": ["warn", { max: 150, skipBlankLines: true, skipComments: true }],
+    },
   },
 ]);
